requirements.txt
mariadb
python-dotenv
requests
PyGithub
pytest
pydantic